saner

+++ Misc/Cloud/setup_apptainer_colab.sh
@@ -5,7 +5,7 @@
 #Install
 export DEBIAN_FRONTEND="noninteractive"
 echo 'debconf debconf/frontend select Noninteractive' | debconf-set-selections
-command -v apt-fast &>/dev/null || echo -e "\[X] FATAL: apt-fast is NOT Installed\n$(exit 1)"
+command -v apt-fast &>/dev/null || echo -e "\n[X] FATAL: apt-fast is NOT Installed\n$(exit 1)"
 BASE_PKGS=(bash binutils coreutils cryptsetup curl fakeroot findutils file g++ git grep jq libc-dev libseccomp-dev moreutils patchelf rsync sed strace tar tree tzdata xz-utils zstd)
 for pkg in "${BASE_PKGS[@]}"; do DEBIAN_FRONTEND="noninteractive" sudo apt-fast install "${pkg}" -y --ignore-missing 2>/dev/null; done
 sudo apt --fix-broken install
@@ -24,7 +24,6 @@ pushd &>/dev/null
  done
  sudo chmod -v "a+x" "./apptainer.deb"
  sudo dpkg -i "./apptainer.deb" || sudo apt --fix-broken install && sudo dpkg -i "./apptainer.deb"
- command -v apptainer &>/dev/null || echo -e "\[X] FATAL: apptainer is NOT Installed\n$(exit 1)"
  sudo apt autoremove -y -qq
  sudo apt autoclean -y -qq
 popd &>/dev/null
@@ -32,16 +31,21 @@ popd &>/dev/null
 
 #-------------------------------------------------------#
 #Pull & Prep Containers
-sudo rm -rf "/containers" 2>/dev/null
-sudo mkdir -pv "/containers" && sudo chown -R "colab:colab" "/containers"
-sudo chmod -R 755 "/containers"
-sudo -u "colab" apptainer build --disable-cache --fix-perms --force --sandbox "/containers/alpine.sif" "docker://docker.io/alpine:edge"
-du -sh "/containers/alpine.sif"
-sudo -u "colab" apptainer build --disable-cache --fix-perms --force --sandbox "/containers/alpine-builder.sif" "docker://ghcr.io/pkgforge/devscripts/alpine-builder:latest"
-du -sh "/containers/alpine-builder.sif"
-sudo -u "colab" apptainer build --disable-cache --fix-perms --force --sandbox "/containers/debian.sif" "docker://docker.io/debian:latest"
-du -sh "/containers/debian.sif"
-sudo -u "colab" apptainer build --disable-cache --fix-perms --force --sandbox "/containers/debian-builder-unstable.sif" "docker://ghcr.io/pkgforge/devscripts/debian-builder-unstable:$(uname -m)"
-du -sh "/containers/debian-builder-unstable.sif"
-ls -lah "/containers"
+if command -v apptainer &> /dev/null; then
+  sudo rm -rf "/containers" 2>/dev/null
+  sudo mkdir -pv "/containers" && sudo chown -R "colab:colab" "/containers"
+  sudo chmod -R 755 "/containers"
+  sudo -u "colab" apptainer build --disable-cache --fix-perms --force --sandbox "/containers/alpine.sif" "docker://docker.io/alpine:edge"
+  du -sh "/containers/alpine.sif"
+  sudo -u "colab" apptainer build --disable-cache --fix-perms --force --sandbox "/containers/alpine-builder.sif" "docker://ghcr.io/pkgforge/devscripts/alpine-builder:latest"
+  du -sh "/containers/alpine-builder.sif"
+  sudo -u "colab" apptainer build --disable-cache --fix-perms --force --sandbox "/containers/debian.sif" "docker://docker.io/debian:latest"
+  du -sh "/containers/debian.sif"
+  sudo -u "colab" apptainer build --disable-cache --fix-perms --force --sandbox "/containers/debian-builder-unstable.sif" "docker://ghcr.io/pkgforge/devscripts/debian-builder-unstable:$(uname -m)"
+  du -sh "/containers/debian-builder-unstable.sif"
+  ls -lah "/containers"
+else
+  echo -e "\n[X] FATAL: apptainer is NOT Installed\n"
+ exit 1 
+fi
 #-------------------------------------------------------#
